+++ README.md
@@ -18,9 +18,14 @@
     - Designed basic login screen with Compose
     - Added user state handling for login success and failure
     - Implemented ViewModel for managing login logic
-  - **Application Dependency Management** :
+  - **Application Dependency Management** - [commit 6876356](https://github.com/ld5ehom/sns-android/commit/68763560103812d030eeb35c74f34f5874ddfb77)
     - Dependency Container Creation: An AppContainer class has been developed to act as a central repository for all dependency injections. This container manages the lifecycle and provision of various services and data sources required across the application.
     - Container Placement in App Class: The container is instantiated and stored within the App class. Placing the container in the App class makes it globally accessible to all other components of the application. This strategy ensures that dependencies are managed centrally and are consistently accessible.
+  - **Refactoring Data Source References**
+    - Transitioned from localDataSource to userLocalDataSource : This change isolates data management to specific views, thereby decluttering activities and improving code clarity.
+  - **Dependency Management with LoginContainer**
+    - Implemented a new LoginContainer : Focuses on managing dependencies exclusively for login functionalities, which enhances resource efficiency and streamlines dependency management.
+
 
 -----
 ## Progress Tracking
